Ajoute de tout les modeles, DAO, Services , Ajout de la configuration des equipement de l immeubles

+++ src/main/java/sn/kd/immoplus/service/UserServiceImpl.java
@@ -16,8 +16,8 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findById(Long id) {
-        return userDAO.findById(id);
+    public User findById(int id) {
+        return userDAO.findById((long) id);
     }
 
     @Override
